refactor: ApplicationPeriodField 분리
